apps/order-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
boto3==1.35.86
httpx==0.28.1
pydantic==2.10.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
prometheus-fastapi-instrumentator==7.0.2
pytest==8.3.4
pytest-asyncio==0.25.0
